backend/requirements.txt
azure-functions==1.21.3
azure-cosmos==4.7.0
azure-storage-blob==12.23.1
azure-cognitiveservices-vision-computervision==0.9.0
msrest==0.7.1
Pillow==11.0.0
pytest==8.3.3
